[reafactor]:pr review 반영 커밋
